backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# 任务队列
celery==5.3.4
flower==2.0.1

# 数据处理
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10

# NLP和AI
openai==1.3.7
transformers==4.35.2
torch==2.1.1
jieba==0.42.1
textrank4zh==0.3

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 配置和验证
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 日志和监控
python-json-logger==2.0.7
prometheus-client==0.19.0

# 时间处理
python-dateutil==2.8.2
pytz==2023.3.post1

# 邮件
aiosmtplib==3.0.1
email-validator==2.1.0.post1